Release v3.0.0
### The "Precision Architecture" Update

A major internal overhaul focused on accuracy, performance, and long-term maintainability.

### Improved
- **Faster deep searches**: The engine now reuses work across refinement tiers, so complex items like books converge to precise results much faster.
- **Better book accuracy**: Fixed several edge cases in book enchantment probability tracking that could cause small inaccuracies at high precision.
- **Performance**: Significant speedups across the board Γאפ eliminated bottlenecks in probability redistribution and cache key handling.

### Fixed
- **Frost Walker** version availability corrected.
- **Quick Charge III** level range corrected.
- **Enchantment conflicts** are now validated automatically instead of relying on manually maintained lists.
- Fixed a rare issue where the search could double-count probability mass near queue capacity.

### Changed
- Complete internal restructuring of the engine, type system, and registry. The public-facing behavior is unchanged, but the codebase is now fully type-safe and modular.
- Expanded test suite from ~30 to 242 tests covering engine correctness, edge cases, and cross-version regression.

+++ src/types/serialization.ts
@@ -2,7 +2,7 @@
  * Compact representation of calculation statistics for efficient transfer.
  */
 export interface CompactStats {
-    comboKeys: BigUint64Array;
+    comboKeys: Float64Array;
     comboProbs: Float64Array;
     rankKeys: Uint32Array;
     rankProbs: Float64Array;
